Split UDP socket tests
Move UDP-specific integration tests into a dedicated module so the socket test layout mirrors the TCP and UDP implementation modules.


Copilot-Session: b5a1a347-37a8-4246-8bbc-306590921475
